[#154547794] reorganize admin
