+++ models/country.py
@@ -1,6 +1,5 @@
-from django.db import models
 from django.core.validators import RegexValidator
-
+from django.db import models
 
 __all__ = [
     'Country',
