+++ models/util.py
@@ -1,0 +1,12 @@
+from django.db import OperationalError
+
+
+def LatestEvent():
+    from tracker.models import Event
+
+    if Event.objects.exists():
+        try:
+            return Event.objects.latest()
+        except (Event.DoesNotExist, OperationalError):
+            return None
+    return None
