+++ templates/admin/automail_prize_contributors.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% load donation_tags %}
 {% load i18n %}
-
+{% load staticfiles %}
 
 {% block title %}Automail Prize Contributors{% endblock %}
